tools/ipc: use an int instead of a ssize_t to store the send()/recv() return
On Windows, send() and recv() return an int, whereas on Unix systems
they return a ssize_t (which is not supported on Windows).

Using an int to store the return of send() and recv() is fine for our
case as we are only reading or writing packets that cannot exceed the
limits of an int.

This makes the code compatible with MSVC.

+++ ngl-tools/ipc.c
@@ -176,24 +176,24 @@ void ipc_pkt_freep(struct ipc_pkt **pktp)
 
 int ipc_send(int fd, const struct ipc_pkt *pkt)
 {
-    const ssize_t n = send(fd, pkt->data, pkt->size, 0);
+    const int n = send(fd, pkt->data, pkt->size, 0);
     if (n < 0) {
         perror("send");
         return NGL_ERROR_IO;
     }
     // XXX: should we loop instead?
     if (n != pkt->size) {
-        fprintf(stderr, "unable write packet (%zd/%d sent)\n", n, pkt->size);
+        fprintf(stderr, "unable write packet (%d/%d sent)\n", n, pkt->size);
         return NGL_ERROR_IO;
     }
     return 0;
 }
 
-static int readbuf(int fd, uint8_t *buf, ssize_t size)
+static int readbuf(int fd, uint8_t *buf, int size)
 {
-    ssize_t nr = 0;
+    int nr = 0;
     while (nr != size) {
-        const ssize_t n = recv(fd, buf + nr, size - nr, 0);
+        const int n = recv(fd, buf + nr, size - nr, 0);
         if (n == 0)
             return 0;
         if (n < 0) {
